fixes #184: Split KafkaEventSinkIT
