+++ consumer/src/main/kotlin/streams/kafka/KafkaEventSink.kt
@@ -2,22 +2,12 @@ package streams.kafka
 
 import kotlinx.coroutines.*
 import org.apache.kafka.clients.consumer.ConsumerConfig
-import org.apache.kafka.clients.consumer.ConsumerRecord
-import org.apache.kafka.clients.consumer.KafkaConsumer
-import org.apache.kafka.clients.consumer.OffsetAndMetadata
-import org.apache.kafka.common.TopicPartition
 import org.neo4j.kernel.configuration.Config
 import org.neo4j.kernel.internal.GraphDatabaseAPI
 import org.neo4j.logging.Log
 import streams.*
-import streams.extensions.offsetAndMetadata
-import streams.extensions.toStreamsSinkEntity
-import streams.extensions.topicPartition
-import streams.service.StreamsSinkEntity
-import streams.service.dlq.DLQData
 import streams.service.dlq.KafkaDLQService
 import streams.utils.Neo4jUtils
-import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.TimeUnit
 
 
@@ -30,11 +20,11 @@ class KafkaEventSink(private val config: Config,
     private lateinit var eventConsumer: StreamsEventConsumer
     private lateinit var job: Job
 
-    private val streamsConfigMap = config.raw.filterKeys {
+    override val streamsConfigMap = config.raw.filterKeys {
         it.startsWith("kafka.") || (it.startsWith("streams.") && !it.startsWith("streams.sink.topic.cypher."))
     }.toMap()
 
-    private val mappingKeys = mapOf(
+    override val mappingKeys = mapOf(
             "zookeeper" to "kafka.zookeeper.connect",
             "broker" to "kafka.${ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG}",
             "from" to "kafka.${ConsumerConfig.AUTO_OFFSET_RESET_CONFIG}",
@@ -144,196 +134,3 @@ class KafkaEventSink(private val config: Config,
 
 }
 
-data class KafkaTopicConfig(val commit: Boolean, val topicPartitionsMap: Map<TopicPartition, Long>) {
-    companion object {
-        private fun toTopicPartitionMap(topicConfig: Map<String,
-                List<Map<String, Any>>>): Map<TopicPartition, Long> = topicConfig
-                .flatMap { topicConfigEntry ->
-                    topicConfigEntry.value.map {
-                        val partition = it.getValue("partition").toString().toInt()
-                        val offset = it.getValue("offset").toString().toLong()
-                        TopicPartition(topicConfigEntry.key, partition) to offset
-                    }
-                }
-                .toMap()
-
-        fun fromMap(map: Map<String, Any>): KafkaTopicConfig {
-            val commit = map.getOrDefault("commit", true).toString().toBoolean()
-            val topicPartitionsMap = toTopicPartitionMap(map
-                    .getOrDefault("partitions", emptyMap<String, List<Map<String, Any>>>()) as Map<String, List<Map<String, Any>>>)
-            return KafkaTopicConfig(commit = commit, topicPartitionsMap = topicPartitionsMap)
-        }
-    }
-}
-
-open class KafkaAutoCommitEventConsumer(private val config: KafkaSinkConfiguration,
-                                        private val log: Log,
-                                        private val dlqService: KafkaDLQService?): StreamsEventConsumer(log, dlqService) {
-
-    private var isSeekSet = false
-
-    val consumer = KafkaConsumer<ByteArray, ByteArray>(config.asProperties())
-
-    lateinit var topics: Set<String>
-
-    override fun withTopics(topics: Set<String>): StreamsEventConsumer {
-        this.topics = topics
-        return this
-    }
-
-    override fun start() {
-        if (topics.isEmpty()) {
-            log.info("No topics specified Kafka Consumer will not started")
-            return
-        }
-        this.consumer.subscribe(topics)
-    }
-
-    override fun stop() {
-        consumer.close()
-        dlqService?.close()
-    }
-
-    private fun readSimple(action: (String, List<StreamsSinkEntity>) -> Unit) {
-        val records = consumer.poll(0)
-        this.topics
-                .filter { topic -> records.records(topic).iterator().hasNext() }
-                .map { topic -> topic to records.records(topic) }
-                .forEach { (topic, topicRecords) -> executeAction(action, topic, topicRecords) }
-    }
-
-    fun executeAction(action: (String, List<StreamsSinkEntity>) -> Unit, topic: String,
-                      topicRecords: MutableIterable<ConsumerRecord<ByteArray, ByteArray>>) {
-        try {
-            action(topic, convert(topicRecords))
-        } catch (e: Exception) {
-            topicRecords
-                    .map { DLQData.from(it, e, this::class.java) }
-                    .forEach{ sentToDLQ(it) }
-        }
-    }
-
-    private fun convert(topicRecords: MutableIterable<ConsumerRecord<ByteArray, ByteArray>>): List<StreamsSinkEntity> = topicRecords
-            .map {
-                try {
-                    "ok" to it.toStreamsSinkEntity()
-                } catch (e: Exception) {
-                    "error" to DLQData.from(it, e, this::class.java)
-                }
-            }
-            .groupBy({ it.first }, { it.second })
-            .let {
-                it.getOrDefault("error", emptyList<DLQData>())
-                        .forEach{ sentToDLQ(it as DLQData) }
-                it.getOrDefault("ok", emptyList()) as List<StreamsSinkEntity>
-            }
-
-    private fun sentToDLQ(dlqData: DLQData) {
-        dlqService?.send(config.streamsSinkConfiguration.dlqTopic, dlqData)
-    }
-
-    private fun readFromPartition(config: KafkaTopicConfig, action: (String, List<StreamsSinkEntity>) -> Unit) {
-        setSeek(config.topicPartitionsMap)
-        val records = consumer.poll(0)
-        config.topicPartitionsMap
-                .mapValues { records.records(it.key) }
-                .filterValues { it.isNotEmpty() }
-                .mapKeys { it.key.topic() }
-                .forEach { (topic, topicRecords) -> executeAction(action, topic, topicRecords) }
-    }
-
-    override fun read(action: (String, List<StreamsSinkEntity>) -> Unit) {
-        readSimple(action)
-    }
-
-    override fun read(topicConfig: Map<String, Any>, action: (String, List<StreamsSinkEntity>) -> Unit) {
-        val kafkaTopicConfig = KafkaTopicConfig.fromMap(topicConfig)
-        if (kafkaTopicConfig.topicPartitionsMap.isEmpty()) {
-            readSimple(action)
-        } else {
-            readFromPartition(kafkaTopicConfig, action)
-        }
-    }
-
-    fun setSeek(topicPartitionsMap: Map<TopicPartition, Long>) {
-        if (isSeekSet) {
-            return
-        }
-        isSeekSet = true
-        consumer.poll(0) // dummy call see: https://stackoverflow.com/questions/41008610/kafkaconsumer-0-10-java-api-error-message-no-current-assignment-for-partition
-        topicPartitionsMap.forEach {
-            when (it.value) {
-                -1L -> consumer.seekToBeginning(listOf(it.key))
-                -2L -> consumer.seekToEnd(listOf(it.key))
-                else -> consumer.seek(it.key, it.value)
-            }
-        }
-    }
-}
-
-class KafkaManualCommitEventConsumer(private val config: KafkaSinkConfiguration,
-                                     private val log: Log,
-                                     private val dlqService: KafkaDLQService?): KafkaAutoCommitEventConsumer(config, log, dlqService) {
-
-    private val topicPartitionOffsetMap = ConcurrentHashMap<TopicPartition, OffsetAndMetadata>()
-
-    override fun start() {
-        if (topics.isEmpty()) {
-            log.info("No topics specified Kafka Consumer will not started")
-            return
-        }
-        this.consumer.subscribe(topics)
-    }
-
-    private fun readSimple(action: (String, List<StreamsSinkEntity>) -> Unit): Map<TopicPartition, OffsetAndMetadata> {
-        val records = consumer.poll(0)
-        return this.topics
-                .filter { topic -> records.records(topic).iterator().hasNext() }
-                .map { topic -> topic to records.records(topic) }
-                .map { (topic, topicRecords) ->
-                    executeAction(action, topic, topicRecords)
-                    topicRecords.last()
-                }
-                .map { it.topicPartition() to it.offsetAndMetadata() }
-                .toMap()
-    }
-
-    private fun readFromPartition(kafkaTopicConfig: KafkaTopicConfig,
-                                  action: (String, List<StreamsSinkEntity>) -> Unit): Map<TopicPartition, OffsetAndMetadata> {
-        setSeek(kafkaTopicConfig.topicPartitionsMap)
-        val records = consumer.poll(0)
-        return kafkaTopicConfig.topicPartitionsMap
-                .mapValues { records.records(it.key) }
-                .filterValues { it.isNotEmpty() }
-                .mapKeys { it.key.topic() }
-                .map { (topic, topicRecords) ->
-                    executeAction(action, topic, topicRecords)
-                    topicRecords.last()
-                }
-                .map { it.topicPartition() to it.offsetAndMetadata() }
-                .toMap()
-    }
-
-    private fun commitData(commit: Boolean, topicMap: Map<TopicPartition, OffsetAndMetadata>) {
-        if (commit) {
-            consumer.commitSync(topicMap)
-        }
-    }
-
-    override fun read(action: (String, List<StreamsSinkEntity>) -> Unit) {
-        val topicMap = readSimple(action)
-        topicPartitionOffsetMap += topicMap
-        commitData(true, topicMap)
-    }
-
-    override fun read(topicConfig: Map<String, Any>, action: (String, List<StreamsSinkEntity>) -> Unit) {
-        val kafkaTopicConfig = KafkaTopicConfig.fromMap(topicConfig)
-        val topicMap = if (kafkaTopicConfig.topicPartitionsMap.isEmpty()) {
-            readSimple(action)
-        } else {
-            readFromPartition(kafkaTopicConfig, action)
-        }
-        topicPartitionOffsetMap += topicMap
-        commitData(kafkaTopicConfig.commit, topicMap)
-    }
-}
