+++ consumer/src/test/kotlin/integrations/kafka/KafkaEventSinkBase.kt
@@ -1,0 +1,92 @@
+package integrations.kafka
+
+import org.apache.kafka.clients.consumer.ConsumerConfig
+import org.apache.kafka.clients.consumer.KafkaConsumer
+import org.apache.kafka.clients.producer.KafkaProducer
+import org.apache.kafka.clients.producer.ProducerConfig
+import org.apache.kafka.common.serialization.ByteArrayDeserializer
+import org.apache.kafka.common.serialization.StringDeserializer
+import org.junit.*
+import org.neo4j.graphdb.factory.GraphDatabaseBuilder
+import org.neo4j.kernel.internal.GraphDatabaseAPI
+import org.testcontainers.containers.KafkaContainer
+import streams.utils.StreamsUtils
+import java.util.*
+
+open class KafkaEventSinkBase {
+
+    companion object {
+
+        private var startedFromSuite = true
+
+        @BeforeClass
+        @JvmStatic
+        fun setUpContainer() {
+            if (!KafkaEventSinkSuiteIT.isRunning) {
+                startedFromSuite = false
+                KafkaEventSinkSuiteIT.setUpContainer()
+            }
+        }
+
+        @AfterClass
+        @JvmStatic
+        fun tearDownContainer() {
+            if (!startedFromSuite) {
+                KafkaEventSinkSuiteIT.tearDownContainer()
+            }
+        }
+    }
+
+    lateinit var graphDatabaseBuilder: GraphDatabaseBuilder
+
+    lateinit var db: GraphDatabaseAPI
+
+    lateinit var kafkaProducer: KafkaProducer<String, ByteArray>
+
+    val cypherQueryTemplate = "MERGE (n:Label {id: event.id}) ON CREATE SET n += event.properties"
+
+    // Test data
+    val dataProperties = mapOf("prop1" to "foo", "bar" to 1)
+    val data = mapOf("id" to 1, "properties" to dataProperties)
+
+    @Before
+    fun setUp() {
+        graphDatabaseBuilder = org.neo4j.test.TestGraphDatabaseFactory()
+                .newImpermanentDatabaseBuilder()
+                .setConfig("kafka.bootstrap.servers", KafkaEventSinkSuiteIT.kafka.bootstrapServers)
+                .setConfig("streams.sink.enabled", "true")
+        val kafkaProperties = Properties()
+        kafkaProperties[ProducerConfig.BOOTSTRAP_SERVERS_CONFIG] = KafkaEventSinkSuiteIT.kafka.bootstrapServers
+        kafkaProperties["zookeeper.connect"] = KafkaEventSinkSuiteIT.kafka.envMap["KAFKA_ZOOKEEPER_CONNECT"]
+        kafkaProperties["group.id"] = "neo4j"
+        kafkaProperties[ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG] = org.apache.kafka.common.serialization.StringSerializer::class.java
+        kafkaProperties[ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG] = org.apache.kafka.common.serialization.ByteArraySerializer::class.java
+
+        kafkaProducer = KafkaProducer(kafkaProperties)
+    }
+
+    @After
+    fun tearDown() {
+        db.shutdown()
+        kafkaProducer.close()
+    }
+
+    fun <K, V> createConsumer(keyDeserializer: String = StringDeserializer::class.java.name,
+                               valueDeserializer: String = ByteArrayDeserializer::class.java.name,
+                               vararg topics: String): KafkaConsumer<K, V> {
+        val props = Properties()
+        props[ProducerConfig.BOOTSTRAP_SERVERS_CONFIG] = KafkaEventSinkSuiteIT.kafka.bootstrapServers
+        props["zookeeper.connect"] = KafkaEventSinkSuiteIT.kafka.envMap["KAFKA_ZOOKEEPER_CONNECT"]
+        props["group.id"] = "neo4j"
+        props["enable.auto.commit"] = "true"
+        props[ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG] = keyDeserializer
+        props[ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG] = valueDeserializer
+        props["auto.offset.reset"] = "earliest"
+        val consumer = KafkaConsumer<K, V>(props)
+        if (!topics.isNullOrEmpty()) {
+            consumer.subscribe(topics.toList())
+        }
+        return consumer
+    }
+
+}
