+++ consumer/src/test/kotlin/integrations/kafka/KafkaStreamsSinkProcedures.kt
@@ -1,130 +1,18 @@
-package integrations
+package integrations.kafka
 
 import kotlinx.coroutines.*
-import org.apache.kafka.clients.consumer.ConsumerConfig
-import org.apache.kafka.clients.consumer.KafkaConsumer
-import org.apache.kafka.clients.producer.KafkaProducer
-import org.apache.kafka.clients.producer.ProducerConfig
 import org.apache.kafka.clients.producer.ProducerRecord
 import org.apache.kafka.common.TopicPartition
-import org.apache.kafka.common.serialization.ByteArrayDeserializer
-import org.apache.kafka.common.serialization.ByteArraySerializer
-import org.apache.kafka.common.serialization.StringDeserializer
-import org.apache.kafka.common.serialization.StringSerializer
-import org.junit.*
 import org.junit.Test
-import org.junit.rules.TestName
 import org.neo4j.kernel.impl.proc.Procedures
 import org.neo4j.kernel.internal.GraphDatabaseAPI
-import org.neo4j.test.TestGraphDatabaseFactory
-import org.testcontainers.containers.KafkaContainer
 import streams.procedures.StreamsSinkProcedures
 import streams.serialization.JSONUtils
-import streams.utils.StreamsUtils
 import java.util.*
 import kotlin.test.*
 
 @Suppress("UNCHECKED_CAST", "DEPRECATION")
-class StreamsSinkProceduresIT {
-    companion object {
-        /**
-         * Kafka TestContainers uses Confluent OSS images.
-         * We need to keep in mind which is the right Confluent Platform version for the Kafka version this project uses
-         *
-         * Confluent Platform | Apache Kafka
-         *                    |
-         * 4.0.x	          | 1.0.x
-         * 4.1.x	          | 1.1.x
-         * 5.0.x	          | 2.0.x
-         *
-         * Please see also https://docs.confluent.io/current/installation/versions-interoperability.html#cp-and-apache-kafka-compatibility
-         */
-        private const val confluentPlatformVersion = "4.0.2"
-        @JvmStatic
-        lateinit var kafka: KafkaContainer
-
-        @BeforeClass @JvmStatic
-        fun setUpContainer() {
-            var exists = false
-            StreamsUtils.ignoreExceptions({
-                kafka = KafkaContainer(confluentPlatformVersion)
-                kafka.start()
-                exists = true
-            }, IllegalStateException::class.java)
-            Assume.assumeTrue("Kafka container has to exist", exists)
-            Assume.assumeTrue("Kafka must be running", kafka.isRunning)
-        }
-
-        @AfterClass @JvmStatic
-        fun tearDownContainer() {
-            StreamsUtils.ignoreExceptions({
-                kafka.stop()
-            }, UninitializedPropertyAccessException::class.java)
-        }
-    }
-
-    private lateinit var db: GraphDatabaseAPI
-
-    private val cypherQueryTemplate = "MERGE (n:Label {id: event.id}) ON CREATE SET n += event.properties"
-
-    @Rule
-    @JvmField
-    var testName = TestName()
-
-    private val EXCLUDE_LOAD_TOPIC_METHOD_SUFFIX = "WithNoTopicLoaded"
-    private val EXCLUDE_SINK_METHOD_SUFFIX = "WithSinkDisabled"
-
-    private val kafkaProperties = Properties()
-
-    private lateinit var kafkaProducer: KafkaProducer<String, ByteArray>
-
-    // Test data
-    private val dataProperties = mapOf("prop1" to "foo", "bar" to 1)
-    private val data = mapOf("id" to 1, "properties" to dataProperties)
-
-
-    @Before
-    fun setUp() {
-        var graphDatabaseBuilder = TestGraphDatabaseFactory()
-                .newImpermanentDatabaseBuilder()
-                .setConfig("kafka.bootstrap.servers", kafka.bootstrapServers)
-        if (!testName.methodName.endsWith(EXCLUDE_LOAD_TOPIC_METHOD_SUFFIX)) {
-            graphDatabaseBuilder.setConfig("streams.sink.topic.cypher.shouldWriteCypherQuery", cypherQueryTemplate)
-        }
-        if (testName.methodName.endsWith(EXCLUDE_SINK_METHOD_SUFFIX)) {
-            graphDatabaseBuilder.setConfig("streams.sink.enabled", "false")
-        }
-        db = graphDatabaseBuilder.newGraphDatabase() as GraphDatabaseAPI
-        db.dependencyResolver.resolveDependency(Procedures::class.java)
-                .registerProcedure(StreamsSinkProcedures::class.java, true)
-
-        kafkaProperties[ProducerConfig.BOOTSTRAP_SERVERS_CONFIG] = kafka.bootstrapServers
-        kafkaProperties["zookeeper.connect"] = kafka.envMap["KAFKA_ZOOKEEPER_CONNECT"]
-        kafkaProperties["group.id"] = "neo4j"
-        kafkaProperties[ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG] = StringSerializer::class.java
-        kafkaProperties[ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG] = ByteArraySerializer::class.java
-
-        kafkaProducer = KafkaProducer(kafkaProperties)
-    }
-
-    @After
-    fun tearDown() {
-        db.shutdown()
-        kafkaProducer.close()
-    }
-
-
-    @Test
-    fun shouldConsumeDataFromProcedureWithSinkDisabled() {
-        val topic = "bar"
-        testProcedure(topic)
-    }
-
-    @Test
-    fun shouldConsumeDataFromProcedure() {
-        val topic = "foo"
-        testProcedure(topic)
-    }
+class KafkaStreamsSinkProcedures : KafkaEventSinkBase() {
 
     private fun testProcedure(topic: String) {
         val producerRecord = ProducerRecord(topic, UUID.randomUUID().toString(), JSONUtils.writeValueAsBytes(data))
@@ -140,13 +28,38 @@ class StreamsSinkProceduresIT {
     }
 
     @Test
+    fun shouldConsumeDataFromProcedureWithSinkDisabled() {
+        graphDatabaseBuilder.setConfig("streams.sink.enabled", "false")
+        db = graphDatabaseBuilder.newGraphDatabase() as GraphDatabaseAPI
+        db.dependencyResolver.resolveDependency(Procedures::class.java)
+                .registerProcedure(StreamsSinkProcedures::class.java, true)
+        val topic = "bar"
+        testProcedure(topic)
+    }
+
+    @Test
+    fun shouldConsumeDataFromProcedure() {
+        db = graphDatabaseBuilder.newGraphDatabase() as GraphDatabaseAPI
+        db.dependencyResolver.resolveDependency(Procedures::class.java)
+                .registerProcedure(StreamsSinkProcedures::class.java, true)
+        val topic = "foo"
+        testProcedure(topic)
+    }
+
+    @Test
     fun shouldTimeout() {
+        db = graphDatabaseBuilder.newGraphDatabase() as GraphDatabaseAPI
+        db.dependencyResolver.resolveDependency(Procedures::class.java)
+                .registerProcedure(StreamsSinkProcedures::class.java, true)
         val result = db.execute("CALL streams.consume('foo1', {timeout: 2000}) YIELD event RETURN event")
         assertFalse { result.hasNext() }
     }
 
     @Test
     fun shouldReadArrayOfJson() {
+        db = graphDatabaseBuilder.newGraphDatabase() as GraphDatabaseAPI
+        db.dependencyResolver.resolveDependency(Procedures::class.java)
+                .registerProcedure(StreamsSinkProcedures::class.java, true)
         val topic = "array-topic"
         val list = listOf(data, data)
         val producerRecord = ProducerRecord(topic, UUID.randomUUID().toString(), JSONUtils.writeValueAsBytes(list))
@@ -165,6 +78,9 @@ class StreamsSinkProceduresIT {
 
     @Test
     fun shouldReadSimpleDataType() {
+        db = graphDatabaseBuilder.newGraphDatabase() as GraphDatabaseAPI
+        db.dependencyResolver.resolveDependency(Procedures::class.java)
+                .registerProcedure(StreamsSinkProcedures::class.java, true)
         val topic = "simple-data"
         val simpleInt = 1
         val simpleBoolean = true
@@ -194,6 +110,9 @@ class StreamsSinkProceduresIT {
 
     @Test
     fun shouldReadATopicPartitionStartingFromAnOffset() = runBlocking {
+        db = graphDatabaseBuilder.newGraphDatabase() as GraphDatabaseAPI
+        db.dependencyResolver.resolveDependency(Procedures::class.java)
+                .registerProcedure(StreamsSinkProcedures::class.java, true)
         val topic = "read-from-range"
         val simpleInt = 1
         val partition = 0
@@ -221,6 +140,9 @@ class StreamsSinkProceduresIT {
 
     @Test
     fun shouldReadFromLatest() = runBlocking {
+        db = graphDatabaseBuilder.newGraphDatabase() as GraphDatabaseAPI
+        db.dependencyResolver.resolveDependency(Procedures::class.java)
+                .registerProcedure(StreamsSinkProcedures::class.java, true)
         val topic = "simple-data-from-latest"
         val simpleInt = 1
         val simpleString = "test"
@@ -252,6 +174,9 @@ class StreamsSinkProceduresIT {
 
     @Test
     fun shouldNotCommit() {
+        db = graphDatabaseBuilder.newGraphDatabase() as GraphDatabaseAPI
+        db.dependencyResolver.resolveDependency(Procedures::class.java)
+                .registerProcedure(StreamsSinkProcedures::class.java, true)
         val topic = "simple-data"
         val simpleInt = 1
         val partition = 0
@@ -271,16 +196,8 @@ class StreamsSinkProceduresIT {
         assertTrue { searchResultMap.containsKey("count") }
         assertEquals(1L, searchResultMap["count"])
 
-        val kafkaProperties = Properties()
-        kafkaProperties[ProducerConfig.BOOTSTRAP_SERVERS_CONFIG] = kafka.bootstrapServers
-        kafkaProperties["zookeeper.connect"] = kafka.envMap["KAFKA_ZOOKEEPER_CONNECT"]
-        kafkaProperties[ConsumerConfig.GROUP_ID_CONFIG] = "neo4j"
-        kafkaProperties[ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG] = StringDeserializer::class.java
-        kafkaProperties[ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG] = ByteArrayDeserializer::class.java
-        val kafkaConsumer = KafkaConsumer<String, ByteArray>(kafkaProperties)
+        val kafkaConsumer = createConsumer<String, ByteArray>()
         val offsetAndMetadata = kafkaConsumer.committed(TopicPartition(topic, partition))
         assertNull(offsetAndMetadata)
     }
-
-
 }
